requirements.txt
Django==4.1.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.1
six==1.16.0
Pillow==9.2.0
factory-boy==3.2.1
django-rest-swagger==2.2.0
drf-yasg==1.21.4
psycopg2==2.9.3
pytest==7.1.3
pytest-django==4.5.2